Change Point X and Y to Int32; encode json using protobuf's marshaler
Added a test in server_test.go to test that "0" is returned for snake
body points. In order to confirm the Json that was correctly encoded,
the test decodes the response body and tests expected values. The
response body was cast as a map[string]interface{} because using a
pb.ListGameFramesResponse omits the empty/zero values. That is to say,
the protobuf tools omit empty/zero values on the way in just like they
do on the way out, which makes them hard to test.

What we're reallying trying to verify is the json on the way out, not
how it's cast into protobuf structs; casting as map[string]interface{}
lets us see the raw json before anything is omitted.